Scripts refactor and add gibbon
