+++ bbopy/algorithms/__init__.py
@@ -1,3 +1,3 @@
 from .base import Algorithm, BoTorchAlgorithm, PymooAlgorithm
-from .botorch_algorithms import BAxUS, TURBO, BO, MES
+from .botorch_algorithms import BAxUS, TURBO, BO, MES, GIBBON
 from .pymoo_algorithms.sobj import GeneticAlgorithm, DifferentialEvolution, CMAES, EvolutionaryStrategy
